Move tests into their proper transport. Relax timeouts on the IPC tests.
